fix: make browseros controller singleton
